Change weight lifted from Int to Double

+++ ExerciseTracker/Data Models/StrengthExercise.swift
@@ -14,7 +14,7 @@ final class StrengthExercise: Exercise {
     var exerciseType: String = ""
     var sets: Int = 0
     var reps: Int = 0
-    var weight: Int = 0
+    var weight: Double = 0.0
     var recordedDate: Date = Date.now
 
     init() {
